Executable start script created to remove wp-content defaults
To keep the local environment tidy a script has been created to run after the first initial 'docker-compose up'. This script will identify the wordpress container for this project and remove the default themes and plugins. Only twentytwentyfour theme is kept in the event that a fallback themeis required while developing. A gitignore file has also been created to ignore each of the defaults in the event that a dev forgets to run scripts/start.sh