Assets Provider created to enqueue scripts
A provider class has been created and added to composer.json. This provider has a method to locate the assets manifest.json file and return the path of the asset name passed. An enqueue_Scripts method is called by the ServiceProvder Register method which enqueues the correct fingerprinted file into wp.